requirements.txt
Django==1.10.5
django-easy-select2==1.3.3
django-extensions==1.7.5
httplib2==0.9.2
git+git://github.com/koniiiik/ksp_login.git@f92287cedaa2647f3e57e9308a5e6c2a65c6232d
Markdown==2.6.7
oauth2==1.9.0.post1
oauthlib==2.0.1
psycopg2==2.6.2
PyJWT==1.4.2
python-openid==2.2.5
requests==2.12.5
requests-oauthlib==0.7.0
six==1.10.0
social-auth-app-django==0.1.0
social-auth-core==0.2.1
